feat: 챌린지 배포 step 관련 함수 분리
